Made some changes on the frontend. started work on the rating component: there is still some work to do but it is close

+++ backend/Controllers/ratingsController.js
@@ -1,24 +1,26 @@
 module.exports = {
     addRating: (req, res) => {
         const db = req.app.get('db')
-        let {staffId, projectId} = req.params 
+        let {projectId} = req.params 
+        let {userId} = req.session         
         let {rating} = req.query
-        db.add_rating({staffId, projectId, rating}).then( response => {
+        db.add_rating({userId, projectId, rating}).then( response => {
             res.status(200).send('Added a rating')
         }).catch(err => res.status(500).send('Could not add rating'))
     },
     updateRating: (req, res) => {
         const db = req.app.get('db')
-        let {ratingId} = req.params 
+        let {userId} = req.session 
         let {rating} = req.query
-        db.update_rating({ratingId, rating}).then( response => {
+        db.update_rating({ratingId, rating, userId}).then( response => {
             res.status(200).send('Updated rating')
         }).catch(err => res.status(500).send('Could not update rating'))
     }, 
     deleteRating: (req, res) => {
         const db = req.app.get('db')
         let {ratingId} = req.params 
-        db.delete_rating({ratingId}).then( response => {
+        let {userId} = req.session 
+        db.delete_rating({ratingId, userId}).then( response => {
             res.status(200).send('Deleted rating')
         }).catch(err => res.status(500).send('Could not delete rating'))
     }
