+++ backend/db/delete_rating.sql
@@ -1,2 +1,3 @@
 delete from project_ratings 
 where id = ${ratingId}
+and user_id = ${userId};
